+++ include/libwebsockets/lws-purify.h
@@ -41,7 +41,7 @@
  * possible to do it in-place, ie, with escaped == string
  */
 LWS_VISIBLE LWS_EXTERN const char *
-lws_sql_purify(char *escaped, const char *string, int len);
+lws_sql_purify(char *escaped, const char *string, size_t len);
 
 /**
  * lws_sql_purify_len() - return length of purified version of input string
@@ -93,12 +93,12 @@ lws_filename_purify_inplace(char *filename);
 
 LWS_VISIBLE LWS_EXTERN int
 lws_plat_write_cert(struct lws_vhost *vhost, int is_key, int fd, void *buf,
-			int len);
+			size_t len);
 LWS_VISIBLE LWS_EXTERN int
-lws_plat_write_file(const char *filename, void *buf, int len);
+lws_plat_write_file(const char *filename, void *buf, size_t len);
 
 LWS_VISIBLE LWS_EXTERN int
-lws_plat_read_file(const char *filename, void *buf, int len);
+lws_plat_read_file(const char *filename, void *buf, size_t len);
 
 LWS_VISIBLE LWS_EXTERN int
 lws_plat_recommended_rsa_bits(void);
